refactor: minimise external dependencies for TextFieldDecimalPosition
